refactor: Test and tidy the entire app
* refactor(login): Style with tailwind and fix visual bugs

* refactor(elements): Policy is now a modal and uses tailwind

* fix(client): Information not updating

* fix(elements): Red delete button

* fix(elements): Button disabled state

* fix: Add more currentColor

* chore: Remove old icons

* fix: SimpleChart update with tailwind

* fix: Styling on statistics modal

* fix: Remove old css variables

* refactor: Dark mode function simplified

* refactor: Dark mode default button themes

* refactor(client): Border color for dark mode on user icon

* fix(client): Okta access checkClient

* chore: Add skeleton colors to tailwind config

* refactor(home): Skeleton loading

* refactor(home): Skeleton loading

* fix: No client placeholder

* fix: Ordered and unordered list styling

* refactor: Remove not needed fade

* refactor: Less plan columns

* refactor: Client user home skeletons

* refactor: Remove products skeleton

* refactor: Skeletons for sessions, plans, templates, and portfolio

* refactor: Skeletons for rest of app and removed old component

* fix: Color picker border color in dark mode

* chore: Move all left componenets to generic

* fix: Error pop up backdrop z-index and remove old code

* fix(booking): Copy change

* chore: Remove code sep

* refactor(state): Add generic loading setter and tidy docs.

* refactor(state): Use loading setter

* fix(booking): Form style

* refactor(state): Use loading setter

* refactor(state): Remove endLoading and reuse setLoading

* refactor(state): Fix cancel state and actions

* refactor(state): Disable button state

* fix(booking): Accepting a request

* fix(elements): Modal reponsiveness and new xs size

* fix(elements): Skeleton colours

* fix(sessions): Has sessions logic show elements

* feat(elements): Backdrop on ColorPicker

* docs: update week color mutation

* refactor(state): setData and setDataDeep rework

* fix(sessions): Empty sessions logic

* refactor(sessions): Simply no sessions logic

* refactor(sessions): Partial fix for week

* fix(sessions): Week color update

* docs(sessions): Tidy up

* refactor: Better Vuex getters

* refactor: Client user plan page

* fix: Calendar event navigate

* chore: Auto-assign PR reviewers

+++ src/components/generic/Booking.vue
@@ -33,16 +33,17 @@
                 href="javascript:void(0)"
                 @click="cancelBooking(booking.id, booking.datetime)"
             >
-                <txt class="text-red-700" bold>Cancel</txt>
+                <txt class="text-red-700">Cancel</txt>
             </a>
-            <div v-else>
+            <div class="flex" v-else>
                 <a
                     v-if="booking.status === 'Pending' && !isInThePast(booking)"
                     href="javascript:void(0)"
                     aria-label="Accept booking"
+                    class="mr-2"
                     @click="acceptBookingRequest(booking.id)"
                 >
-                    <txt>Accept</txt>
+                    <txt class="text-green-700">Accept</txt>
                 </a>
                 <a href="javascript:void(0)" @click="cancelBooking">
                     <txt class="text-red-700">
@@ -78,9 +79,7 @@ export default {
     },
     computed: mapState(["claims"]),
     methods: {
-        /**
-         * Determines the colour of the text for booking statuses.
-         */
+        /** Determines the colour of the text for booking statuses. */
         getStatusColor(status) {
             switch (status) {
                 case "Pending" || "Past":
@@ -92,41 +91,8 @@ export default {
             }
         },
 
-        /**
-         * Accepts a booking request made by the client-user.
-         */
+        /** Accepts a booking request made by the client-user. */
         async acceptBookingRequest() {
-            if (
-                await this.$store.dispatch("openConfirmPopUp", {
-                    title: "Are you sure you want to cancel this booking?",
-                    text: "Your trainer will be notified.",
-                })
-            ) {
-                try {
-                    this.$store.commit("setData", {
-                        attr: "dontLeave",
-                        data: true,
-                    });
-                    await this.$store.dispatch("updateBooking", {
-                        id: this.booking.id,
-                        status: "Scheduled",
-                    });
-                    this.$store.dispatch("openResponsePopUp", {
-                        title: "Booking request accepted",
-                        description:
-                            "Your client will be notified of any upcoming bookings that were accepeted.",
-                    });
-                    this.$store.dispatch("endLoading");
-                } catch (e) {
-                    this.$store.dispatch("resolveError", e);
-                }
-            }
-        },
-
-        /**
-         * Cancels a booking.
-         */
-        async cancelBooking() {
             if (
                 await this.$store.dispatch("openConfirmPopUp", {
                     title: "Are you sure you want to accept this booking?",
@@ -134,9 +100,39 @@ export default {
                 })
             ) {
                 try {
-                    this.$store.commit("setData", {
-                        attr: "dontLeave",
-                        data: true,
+                    this.$store.dispatch("setLoading", {
+                        dontLeave: true,
+                    });
+                    await this.$store.dispatch("updateBooking", {
+                        bookingId: this.booking.id,
+                        clientId: this.isTrainer
+                            ? this.$route.params.client_id
+                            : this.claims.client_id_db,
+                        status: "Scheduled",
+                    });
+                    this.$store.dispatch("openResponsePopUp", {
+                        title: "Booking request accepted",
+                        description:
+                            "Your client will be notified of any upcoming bookings that were accepted.",
+                    });
+                    this.$store.dispatch("setLoading", false);
+                } catch (e) {
+                    this.$store.dispatch("resolveError", e);
+                }
+            }
+        },
+
+        /** Cancels a booking. */
+        async cancelBooking() {
+            if (
+                await this.$store.dispatch("openConfirmPopUp", {
+                    title: "Are you sure you want to cancel this booking?",
+                    text: "Your trainer will be notified.",
+                })
+            ) {
+                try {
+                    this.$store.dispatch("setLoading", {
+                        dontLeave: true,
                     });
                     await this.$store.dispatch("deleteBooking", {
                         clientId: this.isTrainer
@@ -150,7 +146,7 @@ export default {
                         title: "Booking cancelled",
                         description: "Your trainer will be notified",
                     });
-                    this.$store.dispatch("endLoading");
+                    this.$store.dispatch("setLoading", false);
                 } catch (e) {
                     this.$store.dispatch("resolveError", e);
                 }
